requirements.txt
# Data fetching - free, no API key needed
yfinance==0.2.51

# Data processing
pandas==2.2.2
numpy==1.26.4

# Email delivery
secure-smtplib==0.1.1

# Config & env
python-dotenv==1.0.1
pyyaml==6.0.1

# Scheduling (lightweight)
schedule==1.2.2

# HTTP requests
requests==2.32.3
